Split into video crate
