+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = [ "crates/collections", "crates/kernel", "xtask" ]
+members = [ "crates/collections", "crates/kernel", "crates/video", "xtask" ]
 resolver = "2"
 
 [workspace.package]
@@ -11,11 +11,16 @@ license = "MIT OR Apache-2.0"
 repository = "https://github.com/gifnksm/elden-analyzer"
 
 [workspace.dependencies]
+custom-debug = { version = "0.6.2", package = "custom_debug_derive" }
 elden-analyzer-collections = { path = "crates/collections" }
 elden-analyzer-kernel = { path = "crates/kernel" }
+elden-analyzer-video ={ path = "crates/video" }
+ffmpeg = { package = "ffmpeg-next", version = "7.1.0", default-features = false, features = ["codec", "format", "software-scaling"] }
+libc = "0.2.167"
 num-rational = { version = "0.4.2", default-features = false, features = ["std"] }
 num-traits = "0.2.19"
 thiserror = "2.0.3"
+tracing = "0.1.41"
 
 [package]
 name = "elden-analyzer"
@@ -51,14 +56,12 @@ pkg-fmt = "tgz"
 chrono = "0.4.38"
 clap = { version = "4.5.21", features = ["derive"] }
 color-eyre = "0.6.3"
-custom-debug = { version = "0.6.2", package = "custom_debug_derive" }
-custom_debug = { version = "0.6.2", package = "custom_debug_derive" }
 elden-analyzer-collections.workspace = true
 elden-analyzer-kernel.workspace = true
+elden-analyzer-video.workspace = true
 # imageproc = { version = "0.25.0", default-features = false, features = ["display-window"] }
 imageproc = { git = "https://github.com/image-rs/imageproc.git", version = "0.26.0", default-features = false, features = ["display-window"] }
 indicatif = "0.17.9"
-libc = "0.2.167"
 lockfree-object-pool = "0.1.6"
 num-rational.workspace = true
 num-traits.workspace = true
@@ -66,16 +69,10 @@ rayon = "1.10.0"
 regex = "1.11.1"
 sdl2 = { version = "0.36", features = ["use-vcpkg"] }
 tesseract-plumbing = { version = "0.11.0", default-features = false }
-tracing = "0.1.41"
+tracing.workspace = true
 tracing-error = "0.2.1"
 tracing-indicatif = "0.3.6"
 tracing-subscriber = { version = "0.3.19", features = ["env-filter", "json"] }
-
-[target.'cfg(windows)'.dependencies]
-ffmpeg = { package = "ffmpeg-next", version = "7.1.0", default-features = false, features = ["codec", "format", "software-scaling", "static"] }
-
-[target.'cfg(not(windows))'.dependencies]
-ffmpeg = { package = "ffmpeg-next", version = "7.1.0", default-features = false, features = ["codec", "format", "software-scaling"] }
 
 [dev-dependencies]
 [build-dependencies]
