+++ crates/video/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "elden-analyzer-video"
+version.workspace = true
+edition.workspace = true
+publish.workspace = true
+rust-version.workspace = true
+license.workspace = true
+repository.workspace = true
+
+[dependencies]
+custom-debug.workspace = true
+elden-analyzer-kernel.workspace = true
+libc.workspace = true
+num-rational.workspace = true
+num-traits.workspace = true
+thiserror.workspace = true
+tracing.workspace = true
+
+[target.'cfg(windows)'.dependencies]
+ffmpeg = { workspace = true, features = ["static"] }
+
+[target.'cfg(not(windows))'.dependencies]
+ffmpeg = { workspace = true }
