+++ crates/video/README.md
@@ -1,0 +1,3 @@
+<!-- cargo-sync-rdme title [[ -->
+# elden-analyzer-video
+<!-- cargo-sync-rdme ]] -->
