+++ src/bin/elden_analyzer/subcommand/find_ui.rs
@@ -1,12 +1,9 @@
 use std::path::PathBuf;
 
 use color_eyre::eyre::{self, OptionExt};
-use elden_analyzer::{
-    components::Components,
-    util::ImageLogger,
-    video_capture::{Frame, VideoCapture},
-};
+use elden_analyzer::{components::Components, util::ImageLogger};
 use elden_analyzer_kernel::types::time::TimestampRange;
+use elden_analyzer_video::capture::{Frame, VideoCapture};
 use tracing::info;
 
 /// Analyze the video files to extract information
