+++ src/lib.rs
@@ -2,6 +2,5 @@ pub mod algorithm;
 pub mod components;
 pub mod image_process;
 pub mod operator;
-pub mod traits;
 pub mod util;
 pub mod video_capture;
