+++ src/operator/detect_component/histogram_based.rs
@@ -1,11 +1,12 @@
 use std::ops::RangeInclusive;
 
 use elden_analyzer_kernel::types::{clip_rect::ClipRect, rect::Rect};
+use elden_analyzer_video::capture::Frame;
 use imageproc::image::{Luma, Pixel as _, Rgb, RgbImage};
 use num_rational::Ratio;
 use num_traits::ToPrimitive as _;
 
-use crate::{util::ImageLogger, video_capture::Frame};
+use crate::{util::ImageLogger, video_capture::FrameExt as _};
 
 #[derive(Debug)]
 pub struct HistogramBasedComponentDetectorBuilder {
